pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mover"
version = "0.1.4"
description = "Official implementation of MoVer: Motion Verification for Motion Graphics Animations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Jiaju Ma", email = "jiajuma@stanford.edu" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "cairosvg",
    "fastapi==0.115.14",
    "groq",
    "Jinja2",
    "numpy",
    "openai",
    "opencv-contrib-python==4.11.0.86",
    "Pillow",
    "playwright",
    "protobuf",
    "PyYAML",
    "treelib",
    "uvicorn==0.35.0",
    "torch",
    "torchvision",
    "ipykernel>=6.29.5",
]

[project.urls]
Homepage = "https://mover-dsl.github.io/"
Repository = "https://github.com/jama1017/MoVer"
"Bug Tracker" = "https://github.com/jama1017/MoVer/issues"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["mover", "mover.*"]
exclude = ["tests*", "mover.composers*"]

[tool.setuptools.package-data]
"mover" = [
    "assets/*",
    "assets/**/*",
    "*/assets/*",
    "*/assets/**/*",
    "**/assets/*",
    "**/assets/**/*",
    "**/*.html",
    "**/*.css",
    "**/*.js",
    "**/*.json",
]

[tool.setuptools.exclude-package-data]
"mover" = ["composers/**"]

[dependency-groups]
dev = [
    "tree-sitter>=0.24.0",
    "tree-sitter-language-pack>=0.8.0",
]

[project.optional-dependencies]
local = [
    "ollama-python",
    "vllm",
]
